Don't hold bare capsules
